move framework trait into its own sui-indexer-alt-framework-store that provides the generic store traits, and people can opt-in using features
implement store trait in sui-pg-db
