+++ crates/sui-field-count/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub use sui_field_count_derive::*;
 
+/// Trait that provides a constant indicating the number of fields in a struct.
 pub trait FieldCount {
     const FIELD_COUNT: usize;
 }
