+++ crates/sui-indexer-alt-framework-store-traits/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "sui-indexer-alt-framework-store-traits"
+version.workspace = true
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+chrono.workspace = true
+scoped-futures.workspace = true
