+++ crates/sui-indexer-alt-framework-store-traits/src/lib.rs
@@ -1,13 +1,10 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub use crate::pipeline::sequential::Handler as SequentialHandler;
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use scoped_futures::ScopedBoxFuture;
 use std::time::Duration;
-
-pub use scoped_futures;
 
 /// Represents a database connection that can be used by the indexer framework to manage watermark
 /// operations, agnostic of the underlying store implementation.
@@ -87,8 +84,6 @@ pub trait Store: Send + Sync + 'static + Clone {
     async fn connect<'c>(&'c self) -> Result<Self::Connection<'c>, anyhow::Error>;
 }
 
-pub type HandlerBatch<H> = <H as SequentialHandler>::Batch;
-
 /// Extends the Store trait with transactional capabilities, to be used within the framework for
 /// atomic or transactional writes.
 #[async_trait]
@@ -140,12 +135,13 @@ pub struct PrunerWatermark {
 }
 
 impl CommitterWatermark {
-    pub(crate) fn timestamp(&self) -> DateTime<Utc> {
+    pub fn timestamp(&self) -> DateTime<Utc> {
         DateTime::from_timestamp_millis(self.timestamp_ms_hi_inclusive as i64).unwrap_or_default()
     }
 
-    #[cfg(test)]
-    pub(crate) fn new_for_testing(checkpoint_hi_inclusive: u64) -> Self {
+    /// Convenience function for testing, instantiates a CommitterWatermark with the given
+    /// `checkpoint_hi_inclusive` and sets all other values to 0.
+    pub fn new_for_testing(checkpoint_hi_inclusive: u64) -> Self {
         CommitterWatermark {
             epoch_hi_inclusive: 0,
             checkpoint_hi_inclusive,
@@ -156,7 +152,7 @@ impl CommitterWatermark {
 }
 
 impl PrunerWatermark {
-    pub(crate) fn wait_for_ms(&self) -> Option<Duration> {
+    pub fn wait_for_ms(&self) -> Option<Duration> {
         (self.wait_for_ms > 0).then(|| Duration::from_millis(self.wait_for_ms))
     }
 
@@ -164,7 +160,7 @@ impl PrunerWatermark {
     /// no more checkpoints to prune, returns `None`. Otherwise, returns a tuple (from,
     /// to_exclusive) where `from` is inclusive and `to_exclusive` is exclusive. Advance the
     /// watermark as well.
-    pub(crate) fn next_chunk(&mut self, size: u64) -> Option<(u64, u64)> {
+    pub fn next_chunk(&mut self, size: u64) -> Option<(u64, u64)> {
         if self.pruner_hi >= self.reader_lo {
             return None;
         }
