+++ crates/sui-indexer-alt-framework/Cargo.toml
@@ -31,13 +31,13 @@ tracing.workspace = true
 tracing-subscriber = { workspace = true, optional = true }
 url.workspace = true
 
-sui-field-count.workspace = true
 sui-indexer-alt-metrics.workspace = true
 sui-pg-db.workspace = true
 sui-sql-macro.workspace = true
 sui-storage.workspace = true
 sui-types.workspace = true
 sui-rpc-api.workspace = true
+sui-indexer-alt-framework-store-traits.workspace = true
 tonic.workspace = true
 
 [dev-dependencies]
