+++ crates/sui-indexer-alt-reader/src/pg_reader.rs
@@ -91,7 +91,7 @@ impl PgReader {
                 Err(Error::PgConnect(anyhow!("Cancelled while connecting to the database")))
             }
 
-            conn = db.connect() => {
+            conn = db.connection() => {
                 Ok(Connection {
                     conn: conn.map_err(Error::PgConnect)?,
                     metrics: self.metrics.clone(),
