+++ crates/sui-indexer-alt/src/handlers/coin_balance_buckets.rs
@@ -7,7 +7,7 @@ use anyhow::{anyhow, bail, Result};
 use diesel::sql_query;
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection, FieldCount},
     pipeline::{concurrent::Handler, Processor},
     types::{
         base_types::{ObjectID, SuiAddress},
@@ -15,7 +15,6 @@ use sui_indexer_alt_framework::{
         object::{Object, Owner},
         TypeTag,
     },
-    FieldCount,
 };
 use sui_indexer_alt_schema::{
     objects::{StoredCoinBalanceBucket, StoredCoinOwnerKind},
@@ -165,7 +164,7 @@ impl Handler for CoinBalanceBuckets {
 
     const PRUNING_REQUIRES_PROCESSED_VALUES: bool = true;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         let values = values
             .iter()
             .map(|v| v.try_into())
@@ -182,7 +181,7 @@ impl Handler for CoinBalanceBuckets {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> anyhow::Result<usize> {
         use sui_indexer_alt_schema::schema::coin_balance_buckets::dsl;
 
@@ -306,6 +305,7 @@ mod tests {
     use super::*;
     use diesel::QueryDsl;
     use sui_indexer_alt_framework::{
+        store::Store,
         types::{
             base_types::{dbg_addr, MoveObjectType, ObjectID, SequenceNumber, SuiAddress},
             digests::TransactionDigest,
@@ -319,7 +319,9 @@ mod tests {
     use sui_protocol_config::ProtocolConfig;
 
     // Get all balance buckets from the database, sorted by object_id and cp_sequence_number.
-    async fn get_all_balance_buckets(conn: &mut Connection<'_>) -> Vec<StoredCoinBalanceBucket> {
+    async fn get_all_balance_buckets<'a, 'conn>(
+        conn: &'a mut DbConnection<'conn>,
+    ) -> Vec<StoredCoinBalanceBucket> {
         coin_balance_buckets::table
             .order_by((
                 coin_balance_buckets::object_id,
