+++ crates/sui-indexer-alt/src/handlers/ev_struct_inst.rs
@@ -7,7 +7,7 @@ use anyhow::{Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection},
     pipeline::{concurrent::Handler, Processor},
     types::full_checkpoint_content::CheckpointData,
 };
@@ -60,7 +60,7 @@ impl Handler for EvStructInst {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(ev_struct_inst::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -72,7 +72,7 @@ impl Handler for EvStructInst {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -91,6 +91,7 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
+        store::Store,
         types::{event::Event, test_checkpoint_data_builder::TestCheckpointDataBuilder},
         Indexer,
     };
@@ -98,7 +99,9 @@ mod tests {
 
     use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
-    async fn get_all_ev_struct_inst(conn: &mut Connection<'_>) -> Result<Vec<StoredEvStructInst>> {
+    async fn get_all_ev_struct_inst(
+        conn: &mut DbConnection<'_>,
+    ) -> Result<Vec<StoredEvStructInst>> {
         let query = ev_struct_inst::table
             .order_by((
                 ev_struct_inst::tx_sequence_number,
