+++ crates/sui-indexer-alt/src/handlers/tx_affected_addresses.rs
@@ -9,7 +9,7 @@ use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use itertools::Itertools;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection},
     pipeline::{concurrent::Handler, Processor},
     types::{full_checkpoint_content::CheckpointData, object::Owner},
 };
@@ -70,7 +70,7 @@ impl Handler for TxAffectedAddresses {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(tx_affected_addresses::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -82,7 +82,7 @@ impl Handler for TxAffectedAddresses {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -101,13 +101,13 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
+        store::Store, types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
 
     use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
-    async fn get_all_tx_affected_addresses(conn: &mut Connection<'_>) -> Result<Vec<i64>> {
+    async fn get_all_tx_affected_addresses(conn: &mut DbConnection<'_>) -> Result<Vec<i64>> {
         Ok(tx_affected_addresses::table
             .select(tx_affected_addresses::tx_sequence_number)
             .order_by(tx_affected_addresses::tx_sequence_number)
