+++ crates/sui-indexer-alt/src/handlers/tx_affected_objects.rs
@@ -8,7 +8,7 @@ use anyhow::Result;
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection},
     pipeline::{concurrent::Handler, Processor},
     types::{effects::TransactionEffectsAPI, full_checkpoint_content::CheckpointData},
 };
@@ -60,7 +60,7 @@ impl Handler for TxAffectedObjects {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(tx_affected_objects::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -72,7 +72,7 @@ impl Handler for TxAffectedObjects {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -91,13 +91,13 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
+        store::Store, types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
 
     use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
-    async fn get_all_tx_affected_objects(conn: &mut Connection<'_>) -> Result<Vec<i64>> {
+    async fn get_all_tx_affected_objects(conn: &mut DbConnection<'_>) -> Result<Vec<i64>> {
         Ok(tx_affected_objects::table
             .select(tx_affected_objects::tx_sequence_number)
             .order_by(tx_affected_objects::tx_sequence_number)
