+++ crates/sui-indexer-alt/src/handlers/tx_balance_changes.rs
@@ -8,7 +8,7 @@ use anyhow::{Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection},
     pipeline::{concurrent::Handler, Processor},
     types::{
         coin::Coin,
@@ -66,7 +66,7 @@ impl Handler for TxBalanceChanges {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(tx_balance_changes::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -78,7 +78,7 @@ impl Handler for TxBalanceChanges {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -131,13 +131,13 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
+        store::Store, types::test_checkpoint_data_builder::TestCheckpointDataBuilder, Indexer,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
 
     use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
-    async fn get_all_tx_balance_changes(conn: &mut Connection<'_>) -> Result<Vec<i64>> {
+    async fn get_all_tx_balance_changes(conn: &mut DbConnection<'_>) -> Result<Vec<i64>> {
         Ok(tx_balance_changes::table
             .select(tx_balance_changes::tx_sequence_number)
             .order_by(tx_balance_changes::tx_sequence_number)
