+++ crates/sui-indexer-alt/src/handlers/tx_calls.rs
@@ -8,7 +8,7 @@ use anyhow::{Ok, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    db::{Connection, Db},
+    db::{Db, DbConnection},
     pipeline::{concurrent::Handler, Processor},
     types::{full_checkpoint_content::CheckpointData, transaction::TransactionDataAPI},
 };
@@ -62,7 +62,7 @@ impl Handler for TxCalls {
     const MIN_EAGER_ROWS: usize = 100;
     const MAX_PENDING_ROWS: usize = 10000;
 
-    async fn commit<'a>(values: &[Self::Value], conn: &mut Connection<'a>) -> Result<usize> {
+    async fn commit<'a>(values: &[Self::Value], conn: &mut DbConnection<'a>) -> Result<usize> {
         Ok(diesel::insert_into(tx_calls::table)
             .values(values)
             .on_conflict_do_nothing()
@@ -74,7 +74,7 @@ impl Handler for TxCalls {
         &self,
         from: u64,
         to_exclusive: u64,
-        conn: &mut Connection<'a>,
+        conn: &mut DbConnection<'a>,
     ) -> Result<usize> {
         let Range {
             start: from_tx,
@@ -92,6 +92,7 @@ mod tests {
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
+        store::Store,
         types::{base_types::ObjectID, test_checkpoint_data_builder::TestCheckpointDataBuilder},
         Indexer,
     };
@@ -99,7 +100,7 @@ mod tests {
 
     use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
 
-    async fn get_all_tx_calls(conn: &mut Connection<'_>) -> Result<Vec<StoredTxCalls>> {
+    async fn get_all_tx_calls(conn: &mut DbConnection<'_>) -> Result<Vec<StoredTxCalls>> {
         Ok(tx_calls::table
             .order_by((
                 tx_calls::tx_sequence_number,
