+++ crates/sui-pg-db/schema.patch
@@ -1,0 +1,7 @@
+diff --git a/crates/sui-pg-db/src/schema.rs b/crates/sui-pg-db/src/schema.rs
+--- a/crates/sui-pg-db/src/schema.rs
++++ b/crates/sui-pg-db/src/schema.rs
+@@ -1 +1,3 @@
++// Copyright (c) Mysten Labs, Inc.
++// SPDX-License-Identifier: Apache-2.0
+ // @generated automatically by Diesel CLI.
